feat: added log status checker
